HPX/NUMA-aware allocators are now enabled at compile time
