+++ include/aggregation_manager.hpp
@@ -556,8 +556,16 @@ public:
       if (buffer_counter <= slice_alloc_counter) {
         constexpr bool manage_content_lifetime = false;
         buffers_in_use = true;
+#ifndef CPPUDDLE_HAVE_HPX_AWARE_ALLOCATORS
+        // Deactivated HPX-aware allocation...
+        const size_t location_id = 0;
+        // not recommended for performance
+#pragma message                                                                \
+"Warning: Running work aggregation without HPX-aware allocators enabled. Performance negatively impacted!"
+#else
         // get prefered location: aka the current hpx threads location
         const size_t location_id = hpx::get_worker_thread_num();
+#endif
         // Get shiny and new buffer that will be shared between all slices
         // Buffer might be recycled from previous allocations by the
         // buffer_recycler...
