+++ include/cuda_buffer_util.hpp
@@ -7,9 +7,6 @@
 #define CUDA_BUFFER_UTIL_HPP
 
 #include "buffer_manager.hpp"
-#ifdef CPPUDDLE_HAVE_HPX
-#include "hpx_buffer_util.hpp"
-#endif
 
 #include <cuda_runtime.h>
 #include <stdexcept>
@@ -106,14 +103,6 @@ using recycle_allocator_cuda_host =
 template <typename T, std::enable_if_t<std::is_trivial<T>::value, int> = 0>
 using recycle_allocator_cuda_device =
     detail::recycle_allocator<T, detail::cuda_device_allocator<T>>;
-#ifdef CPPUDDLE_HAVE_HPX
-template <typename T, std::enable_if_t<std::is_trivial<T>::value, int> = 0>
-using numa_aware_recycle_allocator_cuda_host =
-    detail::numa_aware_aggressive_recycle_allocator<T, detail::cuda_pinned_allocator<T>>;
-template <typename T, std::enable_if_t<std::is_trivial<T>::value, int> = 0>
-using hpx_aware_recycle_allocator_cuda_device =
-    detail::numa_aware_recycle_allocator<T, detail::cuda_device_allocator<T>>;
-#endif
 
 template <typename T, std::enable_if_t<std::is_trivial<T>::value, int> = 0>
 struct cuda_device_buffer {
