+++ include/hip_buffer_util.hpp
@@ -7,9 +7,6 @@
 #define HIP_BUFFER_UTIL_HPP
 
 #include "buffer_manager.hpp"
-#ifdef CPPUDDLE_HAVE_HPX
-#include "hpx_buffer_util.hpp"
-#endif
 
 #include <hip/hip_runtime.h>
 #include <stdexcept>
@@ -112,15 +109,8 @@ using recycle_allocator_hip_host =
 template <typename T, std::enable_if_t<std::is_trivial<T>::value, int> = 0>
 using recycle_allocator_hip_device =
     detail::recycle_allocator<T, detail::hip_device_allocator<T>>;
-#ifdef CPPUDDLE_HAVE_HPX
-template <typename T, std::enable_if_t<std::is_trivial<T>::value, int> = 0>
-using numa_aware_recycle_allocator_hip_host =
-    detail::numa_aware_aggressive_recycle_allocator<T, detail::hip_pinned_allocator<T>>;
-template <typename T, std::enable_if_t<std::is_trivial<T>::value, int> = 0>
-using hpx_aware_recycle_allocator_hip_device =
-    detail::numa_aware_recycle_allocator<T, detail::hip_device_allocator<T>>;
-#endif
 
+// TODO Is this even required? (cuda version should work fine...)
 template <typename T, std::enable_if_t<std::is_trivial<T>::value, int> = 0>
 struct hip_device_buffer {
   size_t gpu_id{0};
